+++ crates/acts/src/store/data/proc.rs
@@ -32,6 +32,9 @@ impl DbCollectionIden for Proc {
     fn indexed_fields() -> &'static [&'static str] {
         &["state", "mid", "timestamp", "start_time", "end_time"]
     }
+    fn ordered_index_fields() -> &'static [&'static str] {
+        &["timestamp", "start_time", "end_time"]
+    }
     fn version() -> i32 {
         0
     }